+++ cli/helpers/mod.rs
@@ -14,11 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
-#[macro_use]
-extern crate thiserror;
-
-pub mod cli;
-pub mod commands;
-pub mod errors;
-pub mod helpers;
-pub mod updater;
+pub mod snarkos;
+pub mod transaction;
