+++ cli/helpers/snarkos/errors.rs
