+++ cli/helpers/snarkos/methods/get_block_hash.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::snarkos::{
+use crate::helpers::snarkos::{
     errors::RpcRequestError,
     methods::SnarkOSRpcMethod,
     objects::{RequestBody, RpcRequest},
@@ -61,7 +61,7 @@ pub async fn get_block_hash(connection: &RpcClient, block_height: u32) -> Result
 #[cfg(test)]
 mod tests {
     pub use super::*;
-    use crate::snarkos::rpc::RpcAuth;
+    use crate::helpers::snarkos::rpc::RpcAuth;
 
     use tokio::runtime::Runtime;
     use wiremock::{
