+++ cli/helpers/snarkos/methods/mod.rs
